moved around some files to make reading easier for me
